refacto(config): migrate from flag to viper
The goal is to be able to configure the exporter from both flags and
YAML config file.

Ultimately we will be able to load configuration from varenv too, which
is useful when using docker-compose.

Important note:

To avoid breaking change on the CLI, we keep golang flags.
We do not migrate to pflag like viper users usually do.

This is because golang flags authorize long name with one or two dashes:
-port 8080 or --port 8080

Pflag follows UNIX convention, so it does not authorize single dash for
long names. The benefit is that pflag supports shortname concatenation:
-a -b -c ==> -abc.