feat: parse state module return
This will mainly be useful for state triggered by a scheduler.

Salt behavior for a state executed by a scheduler:
- success=true as Salt considers it successfully got the state result
- inconsitency: sometimes the global retcode is equal to 0 even if one
  function of the state failed

So, to better know the success of a state executed by a scheduler, we
will look for "result" in the "return" field, and use them to generate
the global success of the state.

This is only applied on the scheduler, as this issue does not seem to be
present when states are executed directly.