fix: failing scheduled job exposed as successful
This is due to a Salt inconsistency.

For a Scheduled Job, sometimes a failing jobs can have either:
- retcode=0
- success=True
- no retcode defined in each function of the state

The most precise way to know the real status is to parse the "result"
field in Event.Return, meaning get the "result" boolean provided in each
function of the state.

For example, if we have the given state:
  dummy test:
    test.nop:
      - name: "nop"

  success:
    test.succeed_with_changes:
      - name: "success"

  failed:
    test.fail_with_changes:
      - name: "failed"

We get the "result", which is a boolean, for the dummy test, success and
failed. If one of them is returns "false", we consider the state failed.